MathType

+++ src/main/java/org/ojalgo/array/PrimitiveZ008.java
@@ -130,12 +130,12 @@ public class PrimitiveZ008 extends PrimitiveArray {
 
     @Override
     protected void add(final int index, final double addend) {
-        data[index] += Math.round(addend);
+        data[index] += (byte) Math.round(addend);
     }
 
     @Override
     protected void add(final int index, final float addend) {
-        data[index] += Math.round(addend);
+        data[index] += (byte) Math.round(addend);
     }
 
     @Override
